<!--- Provide a general summary of your changes in the Title above -->

## Description

This PR greatly improves performance of checking token definitions. There is something about ~100x speedup. For big accounts with few hundreds token this result to about `70ms -> 1.5ms`, for really huge account with 2k tokens it was about `8000ms -> 70ms`

For nerds: `Array.includes` is `O(n)`, but `Set.has` is `O(1)`

## Related Issue

Resolve <!--- link the issue here -->

## Screenshots: